refactor(world): flatten hierarchy and remove tile map component
 - the world always has the same tile map, so this is just another level of indirection that makes it hard to know which instances of which components do what.
 - Remove world-map-entity and move its code into WorldComponent
 - move world/map folder into world/
