+++ src/app/routes/world/features/combat-feature.component.ts
@@ -20,22 +20,22 @@ import { getFixedEncounterById } from 'app/models/game-data/fixed-encounters';
 import * as Immutable from 'immutable';
 import { List } from 'immutable';
 import { map, take } from 'rxjs/operators';
-import { IEnemy } from '../../../../../app/models/base-entity';
-import { AppState } from '../../../../app.model';
-import { NotificationService } from '../../../../components/notification/notification.service';
-import { Point } from '../../../../core';
-import { CombatEncounterAction } from '../../../../models/combat/combat.actions';
-import { CombatEncounter, IZoneMatch } from '../../../../models/combat/combat.model';
-import { Entity } from '../../../../models/entity/entity.model';
+import { AppState } from '../../../app.model';
+import { NotificationService } from '../../../components/notification/notification.service';
+import { Point } from '../../../core';
+import { ITiledObject } from '../../../core/resources/tiled/tiled.model';
+import { IEnemy } from '../../../models/base-entity';
+import { CombatEncounterAction } from '../../../models/combat/combat.actions';
+import { CombatEncounter, IZoneMatch } from '../../../models/combat/combat.model';
+import { Entity } from '../../../models/entity/entity.model';
 import {
   instantiateEntity,
   ITemplateFixedEncounter,
-} from '../../../../models/game-data/game-data.model';
-import { getGameParty } from '../../../../models/selectors';
-import { GameEntityObject } from '../../../../scene/objects/game-entity-object';
-import { PlayerBehaviorComponent } from '../../behaviors/player-behavior';
-import { TiledFeatureComponent, TiledMapFeatureData } from '../map-feature.component';
-
+} from '../../../models/game-data/game-data.model';
+import { getGameParty } from '../../../models/selectors';
+import { GameEntityObject } from '../../../scene/objects/game-entity-object';
+import { PlayerBehaviorComponent } from '../behaviors/player-behavior';
+import { TiledFeatureComponent } from '../map-feature.component';
 /**
  * A map feature that represents a fixed combat encounter.
  *
@@ -51,7 +51,7 @@ export class CombatFeatureComponent extends TiledFeatureComponent {
   party: PlayerBehaviorComponent | null = null;
 
   // @ts-ignore
-  @Input() feature: TiledMapFeatureData;
+  @Input() feature: ITiledObject | null;
 
   constructor(public store: Store<AppState>, public notify: NotificationService) {
     super();
