+++ src/app/routes/world/features/dialog-feature.component.scss
@@ -1,8 +1,8 @@
-@import "../../../../style/variables";
-@import "../../../../style/mixins";
+@import "../../../style/variables";
+@import "../../../style/mixins";
 // Dialog
 // --------------------------------------------------------------------------
-door-feature {
+dialog-feature {
   .dialog {
     @include pow-dialog();
     z-index: $zCanvasGui;
