+++ src/app/routes/world/features/door-feature.component.scss
@@ -1,8 +1,8 @@
-@import "../../../../style/variables";
-@import "../../../../style/mixins";
+@import "../../../style/variables";
+@import "../../../style/mixins";
 // Dialog
 // --------------------------------------------------------------------------
-dialog-feature {
+door-feature {
   .dialog {
     @include pow-dialog();
     z-index: $zCanvasGui;
