+++ src/app/routes/world/features/store-feature.component.scss
@@ -1,5 +1,5 @@
-@import "../../../../style/variables";
-@import "../../../../style/mixins";
+@import "../../../style/variables";
+@import "../../../style/mixins";
 
 .rpg-game-store {
   max-width: 600px;
