+++ src/app/routes/world/features/stores/armors-store-feature.component.ts
