+++ src/app/routes/world/features/stores/magics-store-feature.component.ts
