+++ src/app/routes/world/features/stores/weapons-store-feature.component.ts
