+++ src/app/routes/world/features/temple-feature.component.scss
@@ -1,5 +1,5 @@
-@import "../../../../style/variables";
-@import "../../../../style/mixins";
+@import "../../../style/variables";
+@import "../../../style/mixins";
 
 temple-feature {
   .temple-view {
