--- src/app/routes/world/map/world-map.entity.html
@@ -1,19 +1,0 @@
-<map-feature-input-behavior [scene]="scene" #input></map-feature-input-behavior>
-<map-feature
-  [class]="feature.class"
-  *ngFor="let feature of activeFeatures$ | async"
-  [tiledMap]="resource$ | async"
-  [tileMap]="this"
-  [scene]="scene"
-  (onClose)="player.escapeFeature()"
-  [feature]="feature"
-></map-feature>
-<world-player
-  #player
-  *ngIf="notTraveling$ | async"
-  [map]="self"
-  [scene]="scene"
-  [point]="renderPoint$ | async"
-  [model]="partyLeader$ | async"
-></world-player>
-<ng-content></ng-content>
