+++ src/app/routes/world/world.component.ts
@@ -6,27 +6,41 @@ import {
   HostListener,
   Input,
   OnDestroy,
+  QueryList,
   ViewChild,
+  ViewChildren,
   ViewEncapsulation,
 } from '@angular/core';
 import { Store } from '@ngrx/store';
-import { Subscription } from 'rxjs';
+import { BehaviorSubject, combineLatest, Observable, Subscription } from 'rxjs';
+import { distinctUntilChanged, map } from 'rxjs/operators';
 import { AppState } from '../../app.model';
 import { TileMapPathBehavior } from '../../behaviors/tile-map-path.behavior';
 import { DebugMenuComponent } from '../../components/debug-menu/debug-menu.component';
 import { LoadingService } from '../../components/loading/loading.service';
 import { NotificationService } from '../../components/notification/notification.service';
 import { PartyMenuComponent } from '../../components/party-menu/party-menu.component';
-import { Point, Rect } from '../../core';
+import { Behavior, IPoint, Point, Rect, TiledTMXResource } from '../../core';
 import { NamedMouseElement, PowInput } from '../../core/input';
+import { ITiledObject } from '../../core/resources/tiled/tiled.model';
+import { Entity } from '../../models/entity/entity.model';
+import { GameStateService } from '../../models/game-state/game-state.service';
+import {
+  getGameKeyData,
+  getGameParty,
+  getGamePartyPosition,
+} from '../../models/selectors';
 import { assertTrue } from '../../models/util';
 import { GameFeatureObject } from '../../scene/objects/game-feature-object';
+import { TileObjectRenderer } from '../../scene/render/tile-object-renderer';
 import { Scene } from '../../scene/scene';
 import { SceneView } from '../../scene/scene-view';
+import { TileMap } from '../../scene/tile-map';
 import { GameWorld } from '../../services/game-world';
 import { RPGGame } from '../../services/rpg-game';
 import { PlayerBehaviorComponent } from './behaviors/player-behavior';
-import { WorldMapComponent } from './map/world-map.entity';
+import { MapFeatureComponent } from './map-feature.component';
+import { WorldPlayerComponent } from './world-player.entity';
 
 @Component({
   selector: 'world',
@@ -36,6 +50,7 @@ import { WorldMapComponent } from './map/world-map.entity';
   templateUrl: './world.component.html',
   host: {
     '(window:resize)': '_onResize($event)',
+
     '(click)': '_onClick($event)',
     '[style.color]': 'styleBackground',
   },
@@ -44,11 +59,92 @@ export class WorldComponent extends SceneView implements AfterViewInit, OnDestro
   @ViewChild('worldCanvas') canvasElementRef: ElementRef;
   @ViewChild(PartyMenuComponent) partyMenu: PartyMenuComponent;
   @ViewChild(DebugMenuComponent) debugMenu: DebugMenuComponent;
-
+  @ViewChild(WorldPlayerComponent) player: WorldPlayerComponent;
+  @ViewChildren('input') behaviors: QueryList<Behavior>;
+  @ViewChildren(MapFeatureComponent) mapFeatures: QueryList<MapFeatureComponent>;
   /**
    * Whether to render debug view information
    */
   @Input() debug: boolean = false;
+
+  /** The {@see TiledTMXResource} map at the input url */
+  resource$: Observable<TiledTMXResource> = this.gameStateService.worldMap$.pipe(
+    distinctUntilChanged(),
+    map((result: TiledTMXResource) => {
+      this.map.setMap(result);
+      return result;
+    })
+  );
+
+  /** Features can be derived after a new map resource has been loaded */
+  features$: Observable<ITiledObject[]> = this.resource$.pipe(
+    map(() => {
+      return this.map.features?.objects || [];
+    })
+  );
+
+  /**
+   * Exclude features that are marked as hidden in the game keydata storage.
+   */
+  activeFeatures$: Observable<any> = combineLatest(
+    [this.store.select(getGameKeyData), this.features$],
+    (keyMap: Immutable.Map<string, any>, features: ITiledObject[]) => {
+      const out = features.filter((f: ITiledObject) => {
+        const idKey = !!keyMap.get(f.properties?.id);
+        const hasAfterKey = typeof f.properties?.after !== 'undefined';
+        const hasUntilKey = typeof f.properties?.until !== 'undefined';
+        const afterKey = keyMap.get(f.properties?.after);
+        const untilKey = keyMap.get(f.properties?.until);
+        // This is disabled because its id is found in key store
+        if (idKey === true) {
+          return false;
+        }
+        // The after key is present and satisfied
+        if (hasAfterKey && !afterKey) {
+          return false;
+        }
+        // The until key is present and satisfied
+        if (hasUntilKey && untilKey) {
+          return false;
+        }
+        // Nothing stops this, it's enabled
+        return true;
+      });
+      return out;
+    }
+  );
+
+  notTraveling$: Observable<boolean> = this.loadingService.loading$.pipe(
+    map((loading: boolean) => !loading)
+  );
+
+  /** Observable of the current player position in the world. Keeps renderPoint in sync after each move. */
+  position$: Observable<IPoint> = this.store.select(getGamePartyPosition).pipe(
+    distinctUntilChanged(),
+    map((position: IPoint) => {
+      this.renderPoint = new Point(position);
+      return this.renderPoint;
+    })
+  );
+
+  /** Observable of Entity representing the player-card leader to be rendered in the world */
+  partyLeader$: Observable<Entity | null> = this.store.select(getGameParty).pipe(
+    map((party: Immutable.List<Entity>) => {
+      return party.get(0) || null;
+    })
+  );
+  private _renderPoint$: BehaviorSubject<IPoint> = new BehaviorSubject(
+    this.renderPoint
+  );
+
+  /** Observable of the current player position in the world */
+  renderPoint$: Observable<Point> = combineLatest(
+    [this.store.select(getGamePartyPosition), this._renderPoint$],
+    (point: IPoint, renderPoint: IPoint) => {
+      this.renderPoint.set(point || renderPoint);
+      return this.renderPoint;
+    }
+  );
 
   /**
    * Escape action handler. Escape out of any active feature or menu
@@ -57,8 +153,8 @@ export class WorldComponent extends SceneView implements AfterViewInit, OnDestro
   handleKeyboardEvent(event: KeyboardEvent) {
     if (event.key === 'Escape') {
       // Escape out of any feature the player is currently in
-      if (this.map.player && this.map.player.feature) {
-        this.map.player.escapeFeature();
+      if (this.player && this.player.feature) {
+        this.player.escapeFeature();
       } else if (this.debugMenu.open) {
         // Otherwise toggle the party menu
         this.debugMenu.open = false;
@@ -79,17 +175,19 @@ export class WorldComponent extends SceneView implements AfterViewInit, OnDestro
     }
   }
 
-  @ViewChild(WorldMapComponent) map: WorldMapComponent;
-
+  /** For referencing in template */
+  readonly self: WorldComponent = this;
+  map: TileMap = new TileMap();
+  renderPoint: Point = new Point();
   styleBackground: string = 'rgba(0,0,0,1)';
   mouse: NamedMouseElement | null = null;
   scene: Scene = new Scene();
-
+  objectRenderer: TileObjectRenderer = new TileObjectRenderer();
   private _subscription: Subscription | null = null;
-
   constructor(
     public game: RPGGame,
     public notify: NotificationService,
+    public gameStateService: GameStateService,
     public loadingService: LoadingService,
     public store: Store<AppState>,
     public world: GameWorld
@@ -111,7 +209,7 @@ export class WorldComponent extends SceneView implements AfterViewInit, OnDestro
     setTimeout(() => this._onResize(), 1);
   }
 
-  public _onResize(event?: Event) {
+  _onResize(event?: Event) {
     super._onResize(event);
     // Camera (window bounds)
     if (this.map) {
@@ -140,7 +238,7 @@ export class WorldComponent extends SceneView implements AfterViewInit, OnDestro
     this._subscription.unsubscribe();
   }
 
-  public _onClick(e: MouseEvent) {
+  _onClick(e: MouseEvent) {
     // Ignore clicks that did not originate on the canvas
     if (e.srcElement !== this.canvas) {
       return;
@@ -170,10 +268,20 @@ export class WorldComponent extends SceneView implements AfterViewInit, OnDestro
     super.renderFrame(elapsed);
     // Map renders features
     if (this.map) {
-      this.map.renderFrame(this, elapsed);
+      this.mapFeatures.forEach((mapFeatureComponent: MapFeatureComponent) => {
+        if (mapFeatureComponent.host) {
+          const data = mapFeatureComponent.host;
+          this.objectRenderer.render(
+            data,
+            data.renderPoint || data.point,
+            this,
+            data.meta
+          );
+        }
+      });
       // Player renders self and target paths
-      if (this.map.player) {
-        this.map.player.renderFrame(this, elapsed);
+      if (this.player) {
+        this.player.renderFrame(this, elapsed);
       }
     }
     if (this.debug) {
